Converted snapserver to use advgetopt.
Now snapserver can read from a config file each command-line argument
(and from the environment too) thanks to advgetopt.
Also converted the main snap-server to using a std::shared_ptr for the
singleton instance.
As a note, each child/plugin should also be a shared_ptr, but this is
beyond the scope of this changed. I put in a #pragma message regarding
this, however.
Note that this code compiles but is untested. Fixes forthcoming.