Fixed crash when server detaches, and added init files.
I added a shared_ptr<> around the g_application object, and it crashed
when the server detached (it looks like it was being doubly deleted).
Anyway remarked out that g_appliction object entirely--it looks like Qt
runs fine without it. But more testing is required.

Added /etc/init and friends files to debian.