Improved xmllint and added plugin comments.
Improved the xmllint support so that we can add a "plugins.dtd" file.
If that file exists, it is used in xmllint as the DTD. Added a clean target
as well, so lint files are easy to clean up.

Added more comments in the fight to make the server pointer either shared
or bare. You decide. :-)