Added missing headers and reorganized snapdb.
A few header files were missing from the project, so I added them. This
includes "snapwebsites.h.in". It makes life in qtcreator a little
easier...

I reorganized snapdb.cpp so the massive if/else block is now broken up,
and each field put into its own respective method. This is in
preparation for moving the business rules for display of row names and
columns into a reusable library module.