Update to allow xsl:include from .xsl files and setup the theme in content types.
(it looks like my previous push did not include all my changes!?)
Moved the system templates from the bare theme to the layout plugin.
Moved the layout data files in a sub-directory.
Fixed a possible bug by saving the layout script in a temporary QByteArray before testing the characters of a string.
Fixed the fact that the layout JavaScript may return an empty string.
Fixed the parameters which should have been variables in layouts.