Much enhancements & fixes on the layout support.
Added support for tarballs (extract/generate) as an archiver (should be used to handle layouts at some points).
Added a new column name support in dbutils.
Added code to allow for layout updates (layout::do_layout_updates())
Added a form.js script, starting on actual functionality on forms.
Added a new type "layout-page" so we can find the list of layouts a website is using.
Enhanced forms.
Changed the output of layouts because some data is binary now (images) to show hex. This could be ameliorated though.
Fixed the resources of the form plugin.
Fixed a problem with the last modification date test in the plugin update code.
Fixed some get_start_date() variable types, it is supposed to be int64_t, not uint64_t.
Fixed a link field name from content::page_page to content::page_type (although maybe we could revert the page_page, only it would be necessary everywhere!)
Fixed the boxes, it would look in the Global instead of the Branch for the list of boxes.
Removed the get_plugin() function from the content plugin definition (it was moved here from the path plugin, and back to the path plugin!)
Put rights (permissions) debug output in #ifdef's so we can turn them off easily.
Updated the snaplayout tool so we can upload images and get a last modification date in the layout table.