Enhanced forms so we can have multiple on a page.
Fixed the code so we can add more than one form on a single page.
Properly change password inputs into text when showing the background text.
Some cleanup to the form.js scripts.
Added a name in all the forms so we can theme them (name goes in the class).
Fixed the unique_id (finally!?) of the forms. Now each form as a different ID.
Fixed the core-form.xsl param definitions with variables (for the name, mtime, year).
Added a registration box. Fixed the corresponding CMakeLists.txt file.