Trying to get the Finball Inc. site to work! 8-)
Fixed a bug with adding CSS, it would add them to the "javascripts" row in the files table. Now it is added to the "css" row as expected.
Fixed the snapwebsites filename (i.e. removed the second "lib" in the filename.)
Updated the dbutils to support the new "css" row.
Some cleanup of different source files.
Allows multiple directories to load plugins from.
Added message handler to avoid warnings of unused variables.