As a test, adding debian packaging info for cassview.
What I hope to accomplish here is a way to keep packaging information in
the main git archive. This way, you can update the changelog first, then
push your changes to sf.net. Then launchpad.net will automatically build
the package and be able to upload the changes after the auto-import is done.
The daily build should then be able to work correctly.