Messages plugin cleanup (no "wrong" dependencies).
Fixed the messages plugin exceptions.
Just use doc in add_javascript
Some cleanup of the code.
Added a filtered signal so we can process the final XML document.
Added a get and set value of input tags in form.js.
Added classes to the user messages to know whether it's an error, a warning, or info.
Added a close button image for pop-up windows.
Allow the error box to be placed absolutely and fading in and out.
Added support so box permissions are always checked in "view" mode.