Some cleanup to get some additional forms to work as expected.
Fix to dbutils as the permissions::dynamic is an 8 bit signed number.
Added a way to transform throws in dbutils to a message on the screen (much easier to debug problems.)
Added a user agent name for backends to run ("Snap! Backend"). I think I had that in one other place, now it is global.
Little clean up.
Added the list plugin, not programmed yet though.
Fixed the verify/resend form.
Added missing files for output to compile properly each time.
"Implemented" an optimization when verifying emails and passwords.