Work on the list plugin (in particular, snap_expr)
Print out a message in cmake so I know whether a system is in debug mode or not.
Added snap_expr class to handle C-like expressions with snap_parser.
Fixed the handling of direct recursivity in snap_parser. We still have a problem in case of a multi-rule recursivity (infinite loops.)
Completed rules and choices possibilities (|, =>>, >> operators).
Verify that no plugin adds the special action named "list" (debug only).
Readded a --filename option to the server so we can add a URI for backends.
Different clean ups all over the place.