Fixed the expression parser.
The expr_list cannot optimize early (by adding sub-items immediately in the higher level list) or cases such as (a, b, c + 3) do not work right: c is added twice resulting in: (a, b, c, c + 3). I removed the early optimization!
Fixed the loading so it actually works. (It makes use of an extra root node.)
Fixed the binary operator, functions return bool so we have to use a different set of SFINAE functions.
Gave a name to parameters of functions so it is easier to know what parameter is invalid.
Added costs to all integer and floating point types.
Create content was making use of the wrong branch number when creating the page type link.
Actually added the branch in links because a link is specific to a branch. However, it is not specific to a revision.
Fixed the listing of links to really only list links (using the "links::%" pattern.)
Added a function to delete a specific (*:*) type of link.
Started fixing the implementation of the list backend, we're getting there.
Different clean ups.