Got a first version that works except for the links between list and items (there is another flow in the links implementation).
Fixed location where the list::last_updated is saved (called wrong function).
Fixed the conditional so only the selected expression from the true and false expressions is computed.
Added a protection so the script cannot read cells marked as secure (through a signal).
Added a fix to help with upgrades of list scripts.
Fixed the variables passed to the script and the item key function returns a string now.
Fixed the scripts so they match the data passed to them.
A bit more clean up.