Added the "Admin" menu.
Added a theme to get default lists (i.e. think default <ul>/<li> HTML list.)
Updated the layout theme support so we can make use of several of those functions from other plugins (i.e. the list plugin for now.)
Fixed the time of last update of the default theme so it doesn't get re-installed over and over again.
Fixed the inclusion of nodes of content to avoid having an unused wrapper.
Moved a couple of DOM helper functions to the qdomhelpers.cpp/h.
Print some of the XSLT on errors because otherwise we cannot know which file it is. The XSLT files are expected to have a comment at the top which should be enough for programmers to know which one it is.
Fixed the string_to_date(), we still had some invalid + 3 to the input pointer.
Added support for converting __DATE__ " " __FILE__ strings to a Unix time_t.
Made the permission error callback that does not generate errors a default in the snapserver.cpp/h.
Added functionality to the token structure for the token replace functionality.
Properly prevent user login/register forms once logged in.
Fixed some permissions.
Some more cleanups, added some missing copyright notices.