All error messages now go to the log.
Took out all streams to the std::cerr, and any error, debug or trace
messages go to the log using the appropriate facility as defined in
log.h. This should keep unwanted stuff from going out to the console,
plus should make sure we get everything we need in the log files on a
working system.