Added code so we can create new pages. At this point you can create a new draft and edit it.
Fixed other copy-of's using node() instead of *.
Added catches for std::exception in case an exception that is not defined from snap_exception is raised.
Reduced the number of frames to show in backtraces to 20. We may want to have it as a parameter in the .conf.
Commented out the stderr version since the logger already do that anyway.
Fixed several bugs in the set_uri() in regard to query strings.
Marked -f and -l parameters as required (this does not mean the option is required on the command line, just that if used, -f and -l must be followed by an argument.)
Added permissions and necessary pages to handle drafts.
Added code so the meta tag description shows up.
Added a flag to know whether a page can ever be deleted (all administration pages cannot be deleted by end users, period.)
Add the Edit button from the editor.js, still need to make it look better and place it in a better location than over the data. Also may want to use some data-blah="" for info about what is being edited.
Upgrade jquery to version 1.11.0.
Some clean ups.