Added snap_stop_thread class, and changed tcp_client_server::accept() to allow a timeout.
The snap_stop_thread class is a work in progress. It will for running a
monitoring thread in the background, which listens to a different port
than the main port. This is to allow snapsignal to send a STOP command
to the snapserver and shut it down cleanly. Also, NLOG will be supported
to allow the user to close the log file cleanly for use with logrotate.

Part of the above changes require the tcp_client_server::accept() method
to be able to timeout while listening on the main port, to periodically
check to see if the STOP signal was sent.