Modified the process class to support background processes.
The process class now allows the run() method to return immediately. I
added is_running() and kill() methods as well.
Also, created a static method for snapwebsites which allow the user to
easily send a udp ping to an address:port pair.
snapinit is a work in progress: it should now be able to support running
process, restart them if they die, and respond to a UDP ping to
shutdown.