Got the Save/Save Draft buttons working in the draft editor.
Updated the C++ warnings: reduced the strict-overlow to 4 instead of 5, tested with -Wconversion and fixed many errors.
"Fixed" comparison between floating point numbers (== and !=).
Fixed the modified_content() signal so it saves the modified info in the data table as it should be.
Added a copy in the new revision so we can create new revisions for users editing work.
Fixed many problems with the token filter.
Removed some debug code.
Added a cool .js (cloud-zoom.js) and will have to create a plugin for it.
Got the POST from the editor.js to work.