Changed listen thread's "word" to a FIFO.
Now snap_listen_thread has a FIFO, a std::list of word_t types. When a
message is received, it is pushed back onto the list, then when the list
is checked, it is popped off the front. If the list is empty, "Waiting"
is returned indicating that nothing is pending.