Finished up an early version of the save and publish capabilities from a draft.
We can now create a draft and publish it! This includes the transformation of the title to a path in JavaScript and C++.
Fixed some entries in the bare and default layouts.
Fixed the highlighting of fields (token) replaced in a complete span or div tag.
Fixed the action saving it in the ipath instead of snap_child (because each ipath can make use of a different action.)
Moved the copying of a row to the dbutils library.