Allow data driven JS/CSS files in layouts. Also allow many different .css and .js files in the same layout.
Enhanced the editor so we can replace the default Save panel with "our" own button(s).
Fixed the auto-focus which would not save the focused element information properly.
Added fields "Layout" and "Depends" to JS and CSS files. Depends is not yet used (needs to be used in the add_xml() instead of the parameters of nearly the same name)
Added the content::add_css() function which is a parallel to the add_javascript() function.
Some clean up.