Fixes so we can register a user under IE9+.
Started implementing the Save of an inline image in the editor.
Implemented the max. width x height of an image in the editor.js and resize accordingly.
Save image filename in the img tag.
Implemented a fix of the forms. Still need to move that code in a .js and make function calls instead of having all the code inline.
Various clean ups.