Enhancements to allow the editor to take over a page as a form.
Added an external CSS for the editor. We'll want to move the CSS from the JavaScript to that file.
Added support for background text in editor forms.
Added support for labels being clicked to focus an editor form widget.
Added and ameliorated different DOM helper functions.
Allow for the theme CSS to be placed BEFORE the JavaScript code.
Changed the snaplayout tool to support any kind of .xml file.
Fixed the child command for the search / create XML as it would duplicate tags by default which was not the first intend.