Added editor validation on the server side and support for proper AJAX responses.
Added the editor validation code.
Added a data driven way of redirecting the user on save in the editor.
Added support for handling error messages in editor forms; we do not yet present those to the user, but they are properly sent back to the AJAX call.
Clean post values by removing starting and ending blanks, including \<br\> tags.
Test whether the XML document is empty before passing it to QXmlQuery which otherwise generates errors "hard" to debug.
Moved the darkenPage code to a function and its CSS to the style.css so that way it can be reused by any dialog and the Save functionality.
Client JS cleans up the data before sending it so the POST is smaller!
Fixed a bug with testing whether the modified flag is still ON in the editor.
Fixed the error message for QXmlQuery() so we clearly distinguish the XSLT and XML files.
Enhanced messages so we can specify the form/editor widget that generates the message.
Allow for the process_post() to generate data and if so avoid the plugin on_path_execute() from running. This is to support AJAX.
Some cleanup.