Fixes so the style sheets work right.
Fixed the snap:body-attributes template so we can dynamically assign attributes from any plugin. (Although at this point plugins cannot really do anything about it...)
Added the "zordered" class to the messages popup so we can define valid z-index-es.
Made the output.js depend on the jquery-extensions which we started to use.
Added a maxZIndex() function to jQuery which searches for the largest z-index value in a set of selectors.