Allow redirect from any plugin. Enhanced the list plugin.
Added _DEBUG for debugs because that's actually what C/C++ expects.
Added NDEBUG when not in debug.
Added the required flag (asterisk) to the checkmark label.
Support an "active" class on the menu items (list really.)
Simplified the page URI for the XML data.
Fixed the $full_path in our XSLT so it actually is correct.
Fixed permission, in case a row gets deleted it would fail with a throw.
Fixed redirect HTTP code on access denied, we want to use 302 (Found) and not any of the 4xx.
Added a PING from the list front end to the backend and avoid the TRACE when not handling a URI.
The list now passes down the list, box, and main page URI and path.
Some cleanup.