Added support for a Dropdown widget.
Fixed the snap_uri class exceptions.
Offer much better control over the action when accessing an editor form.
Added a pause after we get a PING in the pagelist backend so lists are managed faster.
Added a perms() function to use with pagelist. At this point it fails though.
Various cleanups.