Added support for any kind of popup.
Added the qs_layout::layout/theme to support user overwrite of layouts (especially in popups at this point).
Accept a namespace in a dependency name so we can include CSS data from a JavaScript dependency.
Added a minimal layout in the output plugin, used with the popup functionality.
Moved the editor CSS and JS for the darkenPage to the new popup feature.
Fixed a bug in the search of the latest JavaScript/CSS file, the QMap needed to be checked starting from the end.
Cleanups and documentation.