Now setting SNAP_NO_FORK enables the --nofork option.
Passing the -DSNAP_NO_FORK to cmake will no longer just fail to fork. It
now enables a new command line option, "--nofork", which allows the
developer to turn this feature on and off without having to rebuild each
time. When SNAP_NO_FORK is not defined, then the "--nofork" option is no
longer available. In theory, this should never find its way onto a
production server! There are warnings that are emitted.