Worked on the "Save Data" function to get it to work in the new environment.
Made the buffer to MIME type extensible and extracted it from the editor.
Added a read-only state on line edit widgets.
Changed the select-only into a read-only line edit (since we derive from it).
Moved the save dialog style to the CSS file.
Broken up the Editor Save Dialog in a separate class.
Started some work on using the correct widget (i.e. .snap-editor and .snap-content).
Got more on the "modified" status in the widget where it really belongs.
Offer a really extensible way of massaging the data to be sent to the server.
Added an editor form base class.
Added class inheritance functions, need to make use of them now!
Better implemented the support for drag & drop (the ImageBox actually receives an Image object!)
Moved many more events to the rigth object.