Finished up the editor attach_() conversions.
Added another intermediate class to handle the "Edit" button so we do not have to void it in special cases (it is now the other way around.)
Fixed the class name for the edit button to editor-edit-button instead of editor-tooltip.
Updated the class graph, added the EditorWidgetTypeBase which was missing.
Added a Widget Type class graph to make it clearer because there are quite a few classes in there now (8 to be precise)!
Fixed the inheritance initialization and removed all prototype = { ... } because it is not compatible with proper inheritance.
Added and fixed a lot of documentation.
Moved isEmptyBlock() function to the EditorWidget since that's the only place where it is used.
Moved the casts from Output to snapwebsites directly (makes more sense I think, we may even want to have one more .js file with basic things partaining to snapwebsites.)
Fixed the default assignments of arrays and objects (i.e. prototype.someArray_ = []; is wrong because all objects are going to share that array!)
As a result, toobarButtons_ is a static const array!
Moved all the saving / not saving actions in a single function (i.e. disable buttons, gray background, mark form as currently saving.)
Properly prevent the Save button from being clicked more than once until the save operation is done.
Moved the Dropdown widget type right after the Line-Edit since it derives from it, also finally created the proper initialization code.