Getting there! The objects get created and we can already see values and edit.
Fixed the objects superClass_ parameter definition and use.
Added a hidden type, which is useful to pass values to the XSLT files.
Properly signal new type creations.
Moved the editor session generation to the right place. Each edtor form has a session now.
Fixed widget type name "image" with "image-box" as it was supposed to be.
Finished replacing "on" and "off" for the checkmark with "0" and "1".
Added support for a "select_text" filter function that returns the text of a tag.
Enhanced the casts to include a string with a detailed message.
Added code to filter attributes of all the tags.
Fixed the misnamed "snap-content" with "editor-content".
Some optimizations and cleanups.