Fixed the final bugs and got the dropdown and image to work as expected.
Now we allow posts to be updated with replace_postenv().
Fixed background for image boxes (it shows even when the widget is focused.)
Documented some of the classes as AS3 code definitions of classes.
Fixed several calls to the super class with call() as expected.
Offer a map based on the widget name so we can get EditorForm's from user code.
Fixed some documentation and annotations.
Fixed the modification test, now it works right!
Fixed several functions that would use the wrong 'this'.