Added background thread for monitoring pings to snap_backend. Repairs SNAP-52.
Untested, but compiles, so this is a work-in-progress.
The background thread runner, udp_monitor, will only kick in and work
when a signal is actually set. If no signal is set, the background
thread will start, but do nothing.

Each UDP signal string received is queued up in a string list (with
mutex serialization). This allows the backend process to check the
pending signals at their leisure, and insuring that no signal will be
lost.

If there is an error, the foreground process will call
snap_child::exit(1).