Enhanced the editor, AJAX, user functions.
Extracted the AJAX handler from the editor.js and moved it to server-access.js.
Updated the editor inheritance and references documents accordingly.
Applied a fix to the URI generator to avoid a double slash (//) in the path.
Added support for a target when redirecting after a Save in the editor.
Added support for a custom widget so users can create any widget they want.
Fixed an invalid call to give the focus to an element in editor.js.
Fixed the inherits() function, we have to initialize the constructor.
Removed the inPopup flag which was replaced with a target="..." on the redirect.
Added helper functions to retrieve query string parameters in JavaScript.
Added support for simple popups (i.e. Yes / No questions, etc.)
Added helper functions in the users plugin to access users other than the logged in user.
Added a function to create a random password.
Some cleanup.