Fixed a bug in deleting links.
The users plugin would not call the delete link with the correct object, 3 times.
The links plugin would force the branch to the other side branch instead of using the branch defined in the link on delete.
Fixed the editor so on a Save + redirect we do not remove the darken widget.