Broke out the server-access.js from the output and created a seperated plugin.
Added a signal to let plugins know that we are about to output the result to the client.
Changed the output function so there is only one of them for all types of output.
Fixed the fact that at times the system would send HTML when AJAX (XML) was expected.
Cleaned up (replaced with a throw) the Q_ASSERT() in snap_child.cpp.
Various other cleanups.