Better handle cases where we have to ask the user to re-enter their login info.
Put the proper status of the user in the html, body, and meta tags.
Fixed the set_header() in output_result(). They need to be mark as "every-where".
Fixed the output_result() to only use the output_data variable (and not f_output.buffer().)
Made output_result() a public function so we can call it when generating AJAX answers from non-AJAX answers.
Added the server_access dependencies. We still need to automate that function.
Added a function other plugins can call to know whether we are handling an AJAX request.
Fixed the ajax_redirect() so it does not add more than one tag.
Put a better error in case a '-' is found in a token name from the replace_token() function.
Fixed a couple of constructors and removed the this.constructor = ... since we now use the base() function.
Add the "iframe=true" parameter to IFRAME's opened with the popup class.
Some other cleanups.