Added the attachment plugin to clearly separate them from the standard HTML output.
Renamed the output and attachment plugin SNAP_NAME_CONTENT_..._PLUGIN.
Renamed the COMPRESSED name to include GZIP in the name because we may support other compressions.
Support the case where a gzip'ed attachment is larger than the original.
Added a signal so we can handle attachment and send the compressed or converted version.
Commented out the console.log(). We should try to avoid those when publishing because IE and Chrome don't like it.
Fixed the beforeunload event so it actually works (2 missing returns).
Fixed the beforeunload event test so forms that are not to be saved do not get tested.
Fixed the "tel:" so it doesn't get generated if the phone number is empty.
Added a signal before we call the plugin process_post() function.
Added a name to user AJAX data blocks.
Started on the as2js compiler part.
Some cleanup.