Started working on tests on the as2js library.
Fixed several bugs in the string class (found because of the test.)
Added support for a debug version (CMAKE_BUILD_TYPE set to Debug.)
Removed a few overload functions in the String as they were not required.
Added a missing String function (+= with as_char_t).
Fixed a few things in node.cpp so it compiles in debug mode.
Actually implemented the tests for the string class (reached 80% coverage on first version!)