Finished up the String class test, 100% coverage.
Found a bug in the from_wchar() which would clear the string destroying the original even if the new string was invalid.
Found a bug in the from_wchar() as the default result was not set to any value and it could have returned garbage.
Marked the one line in the whole class that cannot be reached as such so LCOV ignores it in its computations.
Renamed all the parser files so their name starts with parser_... that way we can do a breakage of the compiler functions with the same names.