Changed the rc.cpp implementation to use JSON.
Added support for an environment variable named AS2JS_RC for the resource file.
Renamed the get_path() as get_scripts() because it is the path for the scripts.
Fixed the string implementation since we were supposed to support a nullptr.
Added the corresponding test for the nullptr in String.