Finished up the compiler conversion work. It all compiles.
At this point some parts of the package manager are most certainly broken.
Broke the compiler in many separate parts to make it more manageable.
Moved the handling of the depth and parameter indices to the Node class.
Allow for more node conversions.
Added tests in many places to make sure that the compiler would not crash.
Added a coverage test of the Int64 class.