Added a listener plugin to listen to poll from the client.
Added a path in the content attachment_file object so we know the exact resulting path once we created a file.
Added an editor widget named "Dropped File with Preview". It is not complete yet.
Changed the editor variable POST names to start with an underscore (_) so that way we avoid clashes with client's variables.
Fixed bug in editor: the JavaScript code would assume that there was always a title widget on the page.
Added support to send a FormData via the Server Access plugin. This allows us to upload files.
Added support for a userdata object in the Server Access send function.
Allow the AJAX append data to be called early (before the object is fully created.)
Fixed the closure compile script so it accepts multiple references to the $CLOSURE_COMPILER variable.
Made the Widget type base class derive from the ServerAccessCallbacks so that way widgets can use the server access library.
Various clean ups and documentation enhancements.