Added necessary code to support the generation of a preview from a drag & drop event.
Broken up the code analytics and the JavaScript analysis.
Added a UDP signal for the new images plugin backend.
Updated the sendmail and pagelist.
Updated the dbutils to support new rows & cells.
Changed the threads and backend to support the STOP event out of bounds.
Redesigned the FIFO of the backend to use a thread FIFO with thread conditions to wait on so we do not loop forever, we can fall asleep instead.
Added MIME type icons and support to show them in the editor.
Fixed the generated MIME file which was missing a semi-colon.
Added a Dropped File With Preview in the editor (to show previews of PDFs, etc.)
Added corresponding wait to show when something a dragged and dropped and the system works on it.
Moved the thread_life class to the thread.
Added a jQuery extension to rotate objects (too slow for permanent animations, though.)
Removed some more assert()'s.
Various clean ups.