Implemented the success function of the PDF round trip, which now works.
Finished up the images implementation so it supports loading the "preview" image files of a document.
Fixed a "~" with the intended "!" in the attachment code.
Fixed the "set_real_path()" which would clear the parameters of an ipath. Now the clear keeps those parameters on a set_real_path().
Added a get_parent() and get_child() to the ipath because I often do that sort of a thing.