Finished up on the message tests, started on the stream tests.
Removed functions that I added to the MessageCallback() and did not change the outcome of the tests. Still 3 functions that do not get tested.
Added code sample to test abstract classes. Unfortunately, at this the sample proposed only works in cmake v3.0.0 and we only have 2.8.x.
Applied a fix to the string test which had a hicup because it may generate invalid characters instead of bad encoding.