Added support for enumerations with 6 new headers (enum class not supported yet).
Fixed all the operators that were expected to be const but weren't.
Removed the size_t and time_t types as Apple PPC support is probably not required anymore.
Transformed many flags to uint32_t instead of #define.
Moved the different bool typedef's to the new enumeration headers.
The generated header files make use of SFINAE which require C++11 code. If you have a problem with that sub-template, comment those lines out...
Updated the tests accordingly.
Bumped version to 1.4.0.