Fixed the logger after the last fix as it now creates a copy of itself (before it would improperly use a stack version of a logger instance, fixing that problem created a copy.)
Added support for dropped-any-file-with-preview and dropped-image-with-preview.
Applied a fix to properly detect whether an image based widget is defined or not.
Updated the logger to make use of controlled_vars for all its parameters.
Added the missing log()(std::string const&) function.
Implemented the messages to popup in JavaScript for AJAX errors.
Started using the new controlled_vars headers to make use of the bool enumeration.
Allow for anchors in labels attached to a widget.
Added a list of services in the snapserver.conf file for snapinit to use. Updated snapinit accordingly.
Fixed the size test for images as we changed the widget in our event target.
Updated all the CMakeLists.txt to place system headers in their own specific include_directories(). Removed now non-necessary pragmas.
Emit a new JavaScrpt event, "widgetchange", whenever a dropdown selection is made. This happens even if the widget value was not modified.
Added the thrift library to the snapwebsites library (somehow it started to fail linking!)
Transform the special version flags in uint32_t instead of enum values.
Commented out some console.log() -- we really need a way to detect those when creating a real build.
Apply a fix to the table_model.cpp where a row index was not being checked properly.
Bumped the version to 0.1.54.
Various cleanups.