Made the Node.f_parent pointer a weak pointer so nodes auto-delete themselves as expected.
Made all the enumerations 'enum class'. Changed code accordingly.
Added an option so as to forbid the 'with' statement by default.
Removed the destructor. It is just no logical that f_parent would still be set.
Removed unused functions get_children(), get_label_size(), get_label(idx).
Added a couple of test to make sure that only nodes that can be parent or children are.
Removed NODE_ENTRY which is not used anywhere (even in the old implementation.)
Broke up the node.cpp file in 10 files.
Added some more documentation.
Now check the lock when converting the type of a node.
Fixed the min/max/ror/rol operators as <?, >?, <!, >!.
Fixed a bug in the pragma_option() function where the argument pointer could be used when null.
Fixed a bug in the display() function so we can already see our node trees.