Added support in the snapinit tool for --debug so we can have debug output while running our server and backends.
Added a verification to make sure that all services exist before starting anything.
Removed unused code waiting on the sites table to appear. This is part of the snapinit tool now.
Fixed a bug where the backends would return the wrong exit code.
Fixed a bug in the pagelist which would wait 10 seconds (LIST_PROCESSING_LATENCY) before acknoledging a STOP message.
Fixed the snapinit tool so it sends STOP to all the services, then wait for them to exit.
Allow for the lock removal whenever snap_init gets destroyed so that way it is more likely to always happen as required.