Updated versions of the controlled_vars CMakeLists.txt and the changelog.
Fixed the previous version in the controlled_vars changelog from 1.3.7 to 1.4.0.
Fixed the advgetopt "--verbose" to use "--" instead of the &mdash; character (even though it was valid UTF-8, it was incorrect).
Bumped the version of the libcontrolledvars-dev references to 1.4.1.
Bumped the version of the iplock, although we do not build a debian package for it yet.
Bumped the version of the snapwebsites package to 0.1.55.
Added the minimum version of the snapcmakemodules references in various packages.