Implemented full search for layout boxes and tested the page layout::boxes field and it works.
Fixed the name "content-types" with a dash as defined in the database. Used by the taxonomy and now the layout to search a type for a field.
Added the COMMAND_ELEMENT_TEXT, COMMAND_PATH_ELEMENT, COMMAND_IF_[NOT]_ELEMENT_NULL to the field search capabilities.
Fixed the documentation in several places.