Completed the test of the node classes (100% coverage, a few more bugs fixed).
Renamed the NODE_PARAMETERS_FLAG_... as NODE_PARAM_FLAG_... since they apply to the NODE_PARAM and not the NODE_PARAMETERS.
Renamed the NODE_VAR_FLAG_... as NODE_VARIABLE_FLAG_... since they apply to the NODE_VARIABLE and not the NODE_VAR.
Added support to convert an identifier to a label.
Fixed a bug as the IDENTIFIER ':' grammar would not transform the corresponding IDENTIFIER in a LABEL.
Fixed a bug (at least I would imagine so) where the identifier of a function parameter was not getting saved.
Fixed a bug in the compiler_variable.cpp file where I used expr instead of the variable_node variable to set a couple of VARIABLE flags.
Fixed a bug with the display() function which would not print "attrs" before showing the attributes.
Fixed a bug in the package compilation, a test was inverted in the IMPORT implementation.
Fixed a bug in the variable() function of the parser: the variable name was not being saved.
Fixed the set/get_flag() as NODE_PARAM has its own list of flags and it does not share the variable flags.
Fixed the set/get_flag() as the NODE_VAR has no flag, however, the NODE_VAR_ATTRIBUTES does.
Fixed the set/get_string() as many more nodes need to support a string (I re-verified the code to make sure these were all required.)